Add `cleanup` function for removing stale branches
This adds a helper function that can be used as follows:

```bash
cleanup master
```

or

```bash
cleanup some-base
```

These will remove all branches that are "behind" the provided base
branch (e.g. `master` or `develop`). Currently, the protected branches
are `master`, `develop`, and any `rc-` prefixed branches. Use with
caution, as I'm not 100% sure this will always work as desired.